Big update! Added WalletOwnerManager which should handle objects related to players that hold a Wallet. Wallet allows support for multiple currencies, let's say unlimited. Added Currency and CurrencyBuilder (an in game GUI builder for the currency BlobObject) to be available as example and not only that to be available for any plugin which wishes to use multiple currencies in their gamemode, you should now setup this really quick by yourself! EconomyFactory allows you to quickly setup an ObjectDirector of type Currency and simple WalletOwnerManager's up to more advanced ones which will call synchronous (non asynchronous!) events holding your WalletOwner type of while doing the database calls asynchronously.
In a few words, you should be able to make any type of plugin which is able to be a Vault Economy provider in a matter of 2 minutes if you get comfy with the library. I for example am making a Lumber Simulator which would allow the plugin be the Economy provider!